requirements.txt
python-dotenv 
google-generativeai
requests